[GiskardandRobokudo] Ignore virtual joints when setting joint goal in giskard.
Updated robot_description usage from old version to new version in Giskard.

Added options for disabling monitors and allowing gripper collision in achieve_cartesian_goal in Giskard and added these options to the configurations of the world.

Added a globale variable that indicates if robokudo was found or not to avoid using real detection processs module if Robokudo was not found.